Relocate schemaType
