+++ src/schemaType/BooleanType.ts
