+++ src/schemaType/DocumentArrayType.ts
@@ -1,9 +1,9 @@
 import { cloneDeep, isPlainObject, set as setIn } from 'lodash';
-import Document from '../../Document';
-import type Schema from '../../Schema';
-import type { ForeignKeyDbDefinition } from '../classes/ForeignKeyDbTransformer';
-import type { MvRecord } from '../types';
-import { ensureArray, getFromMvArray } from '../utils';
+import Document from '../Document';
+import type Schema from '../Schema';
+import type { ForeignKeyDbDefinition } from '../shared/classes/ForeignKeyDbTransformer';
+import type { MvRecord } from '../shared/types';
+import { ensureArray, getFromMvArray } from '../shared/utils';
 import BaseSchemaType from './BaseSchemaType';
 
 /** A Document Array Schema Type */
