+++ src/schemaType/EmbeddedType.ts
@@ -1,8 +1,8 @@
 import { cloneDeep, isPlainObject, set as setIn } from 'lodash';
-import Document from '../../Document';
-import type Schema from '../../Schema';
-import type { MvRecord } from '../types';
-import { ensureArray } from '../utils';
+import Document from '../Document';
+import type Schema from '../Schema';
+import type { MvRecord } from '../shared/types';
+import { ensureArray } from '../shared/utils';
 import BaseSchemaType from './BaseSchemaType';
 
 /** Embedded Schema Type */
