+++ src/schemaType/ISOTimeType.ts
@@ -1,29 +1,35 @@
 import moment from 'moment';
-import { ISOCalendarDateFormat, mvEpoch } from '../../constants';
-import { TransformDataError } from '../../errors';
-import { handleTypeValidation } from '../utils';
+import { ISOTimeFormat } from '../constants';
+import { TransformDataError } from '../errors';
+import { handleTypeValidation } from '../shared/utils';
 import BaseDateType from './BaseDateType';
 import type { ScalarTypeConstructorOptions } from './BaseScalarType';
 import type { SchemaTypeDefinitionBase } from './BaseSchemaType';
 
-export interface SchemaTypeDefinitionISOCalendarDate extends SchemaTypeDefinitionBase {
-	type: 'ISOCalendarDate';
+export interface SchemaTypeDefinitionISOTime extends SchemaTypeDefinitionBase {
+	type: 'ISOTime';
+	dbFormat?: 's' | 'ms';
 }
 
-/** ISOCalendarDate Schema Type */
-class ISOCalendarDateType extends BaseDateType {
+/** ISOTime Schema Type */
+class ISOTimeType extends BaseDateType {
+	/** Database time format is in milliseconds */
+	private isDbInMs: boolean;
+
 	public constructor(
-		definition: SchemaTypeDefinitionISOCalendarDate,
+		definition: SchemaTypeDefinitionISOTime,
 		options: ScalarTypeConstructorOptions = {},
 	) {
 		super(definition, options);
+		const { dbFormat = 's' } = definition;
+		this.isDbInMs = dbFormat === 'ms';
 
 		// add validators for this type
 		this.validators.unshift(handleTypeValidation(this.validateType));
 	}
 
 	/**
-	 * Transform mv date data to ISO 8601 approved date format (yyyy-mm-dd)
+	 * Transform mv style time data to ISO 8601 approved time format (HH:mm:ss.SSS)
 	 * @throws {@link TransformDataError} Database value could not be transformed to external format
 	 */
 	public transformFromDb(value: null): null;
@@ -33,18 +39,33 @@ class ISOCalendarDateType extends BaseDateType {
 			return null;
 		}
 		const castValue = Number(value);
-		if (!Number.isInteger(castValue)) {
+		if (!Number.isInteger(castValue) || castValue < 0) {
 			throw new TransformDataError({
 				transformClass: this.constructor.name,
 				transformValue: castValue,
 			});
 		}
 
-		return moment(mvEpoch).add(castValue, 'days').format(ISOCalendarDateFormat);
+		if (castValue > 86400000 || (!this.isDbInMs && castValue > 86400)) {
+			throw new TransformDataError({
+				transformClass: this.constructor.name,
+				transformValue: castValue,
+			});
+		}
+
+		const isoTime = moment().startOf('day');
+
+		if (this.isDbInMs) {
+			isoTime.add(castValue, 'milliseconds');
+		} else {
+			isoTime.add(castValue, 'seconds');
+		}
+
+		return isoTime.format(ISOTimeFormat);
 	}
 
 	/**
-	 * Transform ISO 8601 approved date format (yyyy-mm-dd) to mv date data
+	 * Transform ISO 8601 approved time format (HH:mm:ss.SSS) to mv style time data
 	 * @throws {@link TransformDataError} Value could not be transformed to database format
 	 */
 	public transformToDb(value: null): null;
@@ -61,10 +82,15 @@ class ISOCalendarDateType extends BaseDateType {
 			});
 		}
 
-		return String(moment(value).diff(moment(mvEpoch), 'days'));
+		const startOfDay = moment().startOf('day');
+
+		if (this.isDbInMs) {
+			return String(moment(value, ISOTimeFormat).diff(startOfDay, 'milliseconds'));
+		}
+		return String(moment(value, ISOTimeFormat).diff(startOfDay, 'seconds'));
 	}
 
-	/** ISOCalendarDateType data type validator */
+	/** ISOTimeType data type validator */
 	private validateType = async (value: unknown): Promise<boolean> => {
 		if (value == null) {
 			return Promise.resolve(true);
@@ -74,8 +100,8 @@ class ISOCalendarDateType extends BaseDateType {
 			return Promise.resolve(false);
 		}
 
-		return Promise.resolve(moment(value, ISOCalendarDateFormat).isValid());
+		return Promise.resolve(moment(value, ISOTimeFormat).isValid());
 	};
 }
 
-export default ISOCalendarDateType;
+export default ISOTimeType;
