+++ src/schemaType/__tests__/BaseScalarArrayType.test.ts
@@ -1,5 +1,5 @@
 import { BaseScalarType } from '..';
-import type { MvRecord } from '../../types';
+import type { MvRecord } from '../../shared/types';
 import BaseScalarArrayType from '../BaseScalarArrayType';
 import type { SchemaTypeDefinitionString } from '../StringType';
