+++ src/schemaType/__tests__/BaseScalarType.test.ts
@@ -1,6 +1,6 @@
 import type { SchemaTypeDefinitionScalar } from '..';
-import { InvalidParameterError } from '../../../errors';
-import type { MvRecord } from '../../types';
+import { InvalidParameterError } from '../../errors';
+import type { MvRecord } from '../../shared/types';
 import type { ScalarTypeConstructorOptions } from '../BaseScalarType';
 import BaseScalarType from '../BaseScalarType';
 import type { SchemaTypeDefinitionString } from '../StringType';
