+++ src/schemaType/__tests__/BooleanType.test.ts
