+++ src/schemaType/__tests__/DocumentArrayType.test.ts
@@ -1,8 +1,8 @@
-import Document from '../../../Document';
-import type { SchemaDefinition } from '../../../Schema';
-import Schema from '../../../Schema';
-import type { ForeignKeyDbDefinition } from '../../classes/ForeignKeyDbTransformer';
-import type { MvRecord } from '../../types';
+import Document from '../../Document';
+import type { SchemaDefinition } from '../../Schema';
+import Schema from '../../Schema';
+import type { ForeignKeyDbDefinition } from '../../shared/classes/ForeignKeyDbTransformer';
+import type { MvRecord } from '../../shared/types';
 import DocumentArrayType from '../DocumentArrayType';
 
 describe('cast', () => {
