+++ src/schemaType/__tests__/NumberType.test.ts
@@ -1,4 +1,4 @@
-import { InvalidParameterError, TransformDataError } from '../../../errors';
+import { InvalidParameterError, TransformDataError } from '../../errors';
 import type { SchemaTypeDefinitionNumber } from '../NumberType';
 import NumberType from '../NumberType';
